extract os installer to core, add should_load to packages decorator
